mu-plugins: Group Jetpack module functions together
